third_party/WebKit/LayoutTests/http/tests/security/suborigins/suborigin-valid-names-expected.txt


This is a testharness.js-based test.
FAIL Valid suborigin names Uncaught Error: assert_equals: The parent frame should never be able to get the secret value from the child iframe. expected (string) "" but got (undefined) undefined
Harness: the test ran to completion.
